add main loop and refined logic
