+++ bin/board.rb
@@ -1,35 +1,27 @@
+require "./winning.rb"
+require "./player.rb"
+
 # Board class to store the grid and prints on the screen
-class Board
+class Board < Winning
+  attr_reader :grid
+
   @grid = []
 
   def initialize
     @grid = [1, 2, 3, 4, 5, 6, 7, 8, 9]
   end
 
-  def error_out_range
-    puts "That number is not on the Board!"
-  end
-
-  def error_place_taken
-    puts "That place is already taken!"
-  end
-
-  def message_won(player)
-    puts "Congratulations #{player} you won!"
-  end
-
-  def message_board_full
-    puts "Board full with no winner"
-  end
-
-  def write_grid
-    @grid.each_with_index do |item, index|
-      if ((index + 1) % 3).zero?
-        puts "#{item} |"
-        puts "============="
-      else
-        print " #{item} |"
-      end
+  def change_grid(number, char)
+    if @grid[number - 1].nil? || number.zero?
+      error_out_range
+      return false
+    elsif @grid[number - 1] == "X" || @grid[number - 1] == "O"
+      error_place_taken
+      return false
+    elsif @grid[number - 1].between?(1, 9)
+      @grid[number - 1] = char
+      return true
     end
+    false
   end
 end
