+++ bin/playagain.rb
@@ -1,0 +1,7 @@
+def play_again(input)
+  if %w[Y y].include?(input)
+    true
+  else
+    false
+  end
+end
